refactor workbench performance advisor brief routes
